Split local file related metadata management into own extension in NextcloudFilesDatabaseManager
